LPS-127598 Turning checkout button into a submit button
To properly integrate with Data Engine form hijacking we must make go
through the standard form submission path instead of submiting the form
inside an onClick event.

This is because the "submit" event must be sent for Data Engine to pick
it up from there.

This solution is very similar to what journal does on edit_article.jsp.