LPS-141342 portal-kernel: checks for param that will reliably be present
If all checkboxes on the input_permission permission matrix are unchecked, then
no parameters will be present on the request. The code should also check for
the "inputPermissionsViewRole" parameter because it will always be present on a
page that has the input_permission taglib, and therefore we know that the
absence of the other params means:

- no checkboxes checked -> assign no permissions

rather than

- no input_permission fields -> assign default permissions

See the description of https://issues.liferay.com/browse/LPS-85998 for the
original intent of the logic.