LPS-152431 We need to set these to the plid of the original exported Layout, not the plid of the imported Layout, because the FragmentLinkEntryStagedModelDataHandler EXPECTS that they will be set to the plid of the exported Layout and modify them accordingly.
Setting these to the imported Layout's plid here risks causing the
FragmentLinkEntryStagedModelDataHandler to set these to the incorrect
plid in the event that the imported Layout's plid happens to be
identical to a different exported Layout's plid.

The way the code is currently written, it is guaranteed that the Layout
new primary keys map will already be populated with the mapping for this
Layout at this point. So the FragmentLinkEntryStagedModelDataHandler
will be guaranteed to process the exported Layout's plid correctly,
setting it to imported Layout. So I am confident that this will not
cause regressions in vanilla Liferay.